refactor: 코드래빗 피드백 적용
- 오타 수정
- NPE 방지 코드 적용

+++ apps/commerce-collector/src/main/java/com/loopers/infrastructure/metrics/ProductMetricsMonthlyJpaRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
-public interface ProductMetricsMonthlyJpaRepository extends JpaRepository<ProductMetricsMonthly, Integer> {
+public interface ProductMetricsMonthlyJpaRepository extends JpaRepository<ProductMetricsMonthly, Long> {
 
     @Modifying
     @Query("""
